pre-push: let's respect when there is no crowdin

+++ hooks/pre-push.sh
@@ -1,15 +1,17 @@
 #!/bin/bash
 
 [[ -z "$GIT_DIR" ]] && GIT_DIR=".";
-CHANGES="git status --porcelain";
+CHANGES=$(git status --porcelain);
 if [[ ! -z "$CHANGES" ]];
 then
   echo "==> Stash local changes";
   git stash --quiet -u;
 fi;
 
-echo "==> Check en.json for changes";
-node ./node_modules/crowdin-helper/crowdin-helper.js pre-push
+if [ -f "./node_modules/crowdin-helper/crowdin-helper.js" ]; then
+  echo "==> Check en.json for changes";
+  node ./node_modules/crowdin-helper/crowdin-helper.js pre-push
+fi
 
 echo "==> Check TSLint and AOT errors";
 npm run lint && ng build --aot && echo -e "$All fine!${NC}"
